% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvals.R
\name{bootstrap_pvals.merMod}
\alias{bootstrap_pvals.merMod}
\title{Calculate Bootstrap p-values for fixed effects}
\usage{
bootstrap_pvals.merMod(
  model,
  type,
  B,
  resample = NULL,
  reb_type = NULL,
  hccme = NULL,
  aux.dist = NULL
)
}
\description{
Calculate Bootstrap p-values for fixed effects
}
