% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstrap_nlme.R
\name{case.lme}
\alias{case.lme}
\title{Cases Bootstrap}
\usage{
case.lme(model, fn, B, extra_step = FALSE)
}
\arguments{
\item{extra_step}{add the extra step}
}
\value{
list
}
\description{
The Cases Bootstrap samples entire cases to generate the bootstrap.
}
\details{
add details here
}

