% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstrap_lme4.R
\name{case.lmerMod}
\alias{case.lmerMod}
\title{Cases Bootstrap}
\usage{
case.lmerMod(model, fn, B, extra_step = FALSE)
}
\arguments{
\item{extra_step}{add the extra step}
}
\value{
list
}
\description{
The Cases Bootstrap samples entire cases to generate the bootstrap.
}
\details{
add details
}
\references{
vanderLeeden:208kv
}

