% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstrap_lme4.R
\name{cgr.lmerMod}
\alias{cgr.lmerMod}
\title{CGR Bootstrap}
\usage{
cgr.lmerMod(model, fn, B)
}
\value{
list
}
\description{
CGR Bootstrap
}
\details{
add details later
}
\references{
Carpenter, J. R., Goldstein, H., and Rasbash, J. (2003)
A novel bootstrap procedure for assessing the relationship
between class size and achievement. \emph{Journal of the Royal
Statistical Society. Series C. Applied Statistics}, 52(4), 431--443.
doi:10.1111/1467-9876.00415
}

