% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstrap_lme4.R, R/generics.R
\name{cgr_bootstrap.lmerMod}
\alias{cgr_bootstrap}
\alias{cgr_bootstrap.lmerMod}
\title{CGR Bootstrap}
\usage{
\method{cgr_bootstrap}{lmerMod}(model, fn, B)

cgr_bootstrap(model, fn, B)
}
\arguments{
\item{model}{The model object you wish to bootstrap.}

\item{fn}{A function returning the statistic(s) of interest.}

\item{B}{The number of bootstrap resamples.}
}
\value{
list
}
\description{
CGR Bootstrap
}
\details{
add details later
}
\references{
Carpenter, J. R., Goldstein, H., and Rasbash, J. (2003)
 A novel bootstrap procedure for assessing the relationship
 between class size and achievement. \emph{Journal of the Royal
 Statistical Society. Series C. Applied Statistics}, 52(4), 431--443.
 doi:10.1111/1467-9876.00415
}

