% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_messages.R
\name{collect_warnings}
\alias{collect_warnings}
\title{Collect all the warnings, errors, and messages from model refits}
\usage{
collect_warnings(.x)
}
\description{
Collect all the warnings, errors, and messages from model refits
}
