% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_lmersamp.R
\name{combine_lmeresamp}
\alias{combine_lmeresamp}
\title{combine}
\usage{
combine_lmeresamp(...)
}
\description{
Combines processes split for parallelization.
}
\details{
This helper function combines the different processes split for 
parallelization to yield unified output and bootstrap statistics.
}
