% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_lmersamp.R
\name{combine_pvals}
\alias{combine_pvals}
\title{Combine p-values}
\usage{
combine_pvals(...)
}
\arguments{
\item{...}{two or more summary data frames produced by \code{bootstrap_pvals}.}
}
\description{
Combines bootstrap p-values from processes split for parallelization.
}
\details{
This helper function combines the different summary tables produced by
\code{bootstrap_pvals()} when run in parallel to yield unified output
and a single summary table.
}
