% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_messages.R
\name{factory}
\alias{factory}
\title{Catch errors and warnings}
\usage{
factory(
  fun,
  debug = FALSE,
  errval = "An error occurred in the factory function",
  types = c("message", "warning", "error")
)
}
\arguments{
\item{fun}{The function to be turned into a factory}

\item{debug}{print debugging statements?}

\item{errval}{the value to be returned from the function if an error is thrown}

\item{types}{which types to catch?}
}
\value{
The result of the function given to turn into a factory.  If this function was in error "An error as occurred" as a character element.  factory-error and factory-warning attributes may also be set as appropriate.
}
\description{
This function is modified from a version in lme4, which was based on
a factory written by Martin Morgan on Stack Overflow (see below).  
Factory generates a function which is appropriately wrapped by error handlers.  
If there are no errors and no warnings, the result is provided.  
If there are warnings but no errors, the result is provided with a warn attribute set.
If there are errors, the result returns is a list with the elements of warn and err.
}
\examples{
f.log <- factory(log)
f.log("a")
f.log.NA <- factory(log,errval=NA)
f.log.NA("a")
f.as.numeric <- factory(as.numeric)
f.as.numeric(c("a","b",1))
}
\references{
\url{http://stackoverflow.com/questions/4948361/how-do-i-save-warnings-and-errors-as-output-from-a-function}
}
\author{
Martin Morgan; Modified by Russell S. Pierce and Ben Bolker
}
