% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstrap_nlme.R
\name{parametric.lme}
\alias{parametric.lme}
\title{Parametric Bootstrap}
\usage{
parametric.lme(model, fn, B)
}
\value{
list
}
\description{
The Parametric Bootstrap is uses the parametrically estimated
distribution function of the data to generate bootstrap samples.
}
\details{
This function extracts the fixed effects, simulates from the model, refits the model
and then returns the results in a list.
}

