% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bootstrap_lme4.R
\name{parametric.lmerMod}
\alias{parametric.lmerMod}
\title{Parametric Bootstrap}
\usage{
parametric.lmerMod(model, fn, B)
}
\value{
list
}
\description{
The Parametric Bootstrap is uses the parametrically estimated
distribution function of the data to generate bootstrap samples.
}
\details{
This function extracts the fixed effects, simulates from the model, refits the model
and then returns the results in a list.
}
\references{
Chambers:2013ba
vanderLeeden:208kv
}

