% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4_utility_functions.R
\name{prep_cases.merMod}
\alias{prep_cases.merMod}
\title{Set up for case resampling in lme4}
\usage{
prep_cases.merMod(model, resample, orig_data)
}
\description{
Set up for case resampling in lme4
}
