% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.coef_tbl}
\alias{print.coef_tbl}
\title{Print coefficients from a \code{coef_tbl} object}
\usage{
print.coef_tbl(x)
}
\arguments{
\item{x}{the coef_tbl object to print}
}
\description{
Print table of coefficients produced by \code{bootstrap_pvals}.
}
