% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bootstrap_lme4.R
\name{reb.lmerMod}
\alias{reb.lmerMod}
\title{REB Bootstrap}
\usage{
reb.lmerMod(model, fn, B, reb_type = 0)
}
\arguments{
\item{reb_type}{Chooses the type of REB bootstrap}
}
\value{
list
}
\description{
REB Bootstrap
}
\details{
add details
}
\references{
Chambers, R. and Chandra, H. (2013)
A Random Effect Block Bootstrap for Clustered Data.
\emph{Journal of Computational and Graphical Statistics},
22(2), 452–470. doi:10.1080/10618600.2012.681216
}

