% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstrap_nlme.R
\name{residual.lme}
\alias{residual.lme}
\title{Residual Bootstrap}
\usage{
residual.lme(model, fn, B)
}
\value{
list
}
\description{
The Residual Bootstrap uses residuals to generate bootstrap samples.
}
\details{
This function extracts the Xbetas, random effects, residuals, and Z
design matrix in order to resample the residuals and complete the
bootstrap process.
}

