% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bootstrap_lme4.R
\name{residual.lmerMod}
\alias{residual.lmerMod}
\title{Residual Bootstrap}
\usage{
residual.lmerMod(model, fn, B)
}
\value{
list
}
\description{
The Residual Bootstrap uses residuals to generate bootstrap samples.
}
\details{
Resamples the residuals and complete the bootstrap process.
}
\references{
vanderLeeden:208kv
}

