% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.lmeresamp}
\alias{summary.lmeresamp}
\title{Calculate summary statistics for \code{lmeresamp} objects}
\usage{
\method{summary}{lmeresamp}(object, ...)
}
\arguments{
\item{object}{The lmeresamp object to be summarized.}

\item{...}{not used}
}
\description{
Calculate basic summary statistics such as the mean, standard error, and bias of the
bootstrap replicates.
}
\details{
If the bootstrap statistics are stored in a vector (as opposed to a data frame or tibble),
then summary statistics will be calculated. The printed data frame will include
the name of the term (if applicable), the observed value (\code{observed}), the mean of the bootstrap replicated
(\code{rep.mean}), the standard error (\code{se}), and the bootstrap bias estimate (\code{bias}).
In addition, the number of resamples will be printed. If any messages, warnings, or errors were
generated during the bootstrap procedure, they will be summarized below, and you should check the
\code{message}, \code{warning}, and \code{error} elements of the \code{lmeresamp} object to
investigate further.
}
