% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bootstrap_lme4.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap for LMEs}
\usage{
bootstrap(model, fn, type, B)
}
\arguments{
\item{model}{The original model to use}

\item{fn}{The function the user is interested in testing}

\item{type}{The \code{type} of bootstrap requested, see details for types}

\item{B}{The number of bootstrap simulations}
}
\description{
\code{bootstrap} helps streamline the bootstrap process for the parametric,
residual, cases, CGR, and REB bootstraps.
}
\details{
To choose a bootstrap use the \code{type} parameter.
For parametric use \code{"par"}, residual use \code{"res"}, cases use \code{"case"},
CGR use \code{"cgr"}, and REB use \code{"reb"}. The REB bootstrap has two types
which defaults to 1 but can be chosen using \code{"reb", "reb1", "reb2"}.
}

