% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstrap_lme4.R, R/generics.R
\name{bootstrap.lmerMod}
\alias{bootstrap}
\alias{bootstrap.lmerMod}
\title{Bootstrap Nested LMEs}
\usage{
\method{bootstrap}{lmerMod}(model, fn, type, B, extra_step, reb_type)

bootstrap(model, fn, type, B, replace, reb_type)
}
\arguments{
\item{model}{The model object you wish to bootstrap.}

\item{fn}{A function returning the statistic(s) of interest.}

\item{type}{A character string indicating the type of bootstrap that is being
requested. Possible values are \code{"par"} (parametric), \code{"resid"}
(residual), \code{"case"}, \code{"cgr"} (), or one of the three versions
of the random effect block bootstrap ("reb", "reb1", "reb2").}

\item{B}{The number of bootstrap resamples.}

\item{extra_step}{}

\item{reb_type}{Chooses the type of REB bootstrap}
}
\description{
\code{bootstrap} helps streamline the bootstrap process for the parametric,
residual, cases, CGR, and REB bootstraps.
}
\references{
Carpenter:2003uy
   Chambers:2013ba
   Morris:2002tj
   vanderLeeden:208kv
}

