% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvals.R
\name{bootstrap_pvals}
\alias{bootstrap_pvals}
\alias{bootstrap_pvals.merMod}
\alias{bootstrap_pvals.lme}
\title{Calculate Bootstrap p-values for fixed effects}
\usage{
bootstrap_pvals(
  model,
  type,
  B,
  resample = NULL,
  reb_type = NULL,
  hccme = NULL,
  aux.dist = NULL
)

\method{bootstrap_pvals}{merMod}(
  model,
  type,
  B,
  resample = NULL,
  reb_type = NULL,
  hccme = NULL,
  aux.dist = NULL
)

\method{bootstrap_pvals}{lme}(
  model,
  type,
  B,
  resample = NULL,
  reb_type = NULL,
  hccme = NULL,
  aux.dist = NULL
)
}
\arguments{
\item{model}{The model object you wish to bootstrap.}

\item{type}{A character string indicating the type of bootstrap that is being
requested. Possible values are \code{"parametric"}, \code{"residual"}, 
\code{"case"}, \code{"wild"}, or \code{"reb"} 
(random effect block bootstrap).}

\item{B}{The number of bootstrap resamples.}

\item{resample}{A logical vector specifying whether each level of the model 
should be resampled in the cases bootstrap. The levels should be specified 
from the highest level (largest cluster) of the hierarchy to the lowest 
(observation-level); for example for students within a school, specify the 
school level first, then the student level.}

\item{reb_type}{Specification of what random effect block bootstrap version to
implement. Possible values are \code{0}, \code{1} or \code{2}.}

\item{hccme}{either \code{"hc2"} or \code{"hc3"}, indicating which 
heteroscedasticity consistent covariance matrix estimator to use.}

\item{aux.dist}{one of \code{"mammen"}, \code{"rademacher"}, \code{"norm"}, 
\code{"webb"}, or \code{"gamma"} indicating which auxiliary 
distribution to draw the errors from}
}
\description{
Calculate Bootstrap p-values for fixed effects
}
