% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstrap_lme4.R, R/generics.R
\name{case_bootstrap.lmerMod}
\alias{case_bootstrap}
\alias{case_bootstrap.lmerMod}
\title{Cases Bootstrap}
\usage{
\method{case_bootstrap}{lmerMod}(model, fn, B, replace)

case_bootstrap(model, fn, B, replace)
}
\arguments{
\item{model}{The model object you wish to bootstrap.}

\item{fn}{A function returning the statistic(s) of interest.}

\item{B}{The number of bootstrap resamples.}
}
\value{
list
}
\description{
The Cases Bootstrap samples entire cases to generate the bootstrap.
}
\details{
add details
}
\references{
vanderLeeden:208kv
}

