% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstrap_lme4.R, R/bootstrap_nlme.R, R/generics.R
\name{case_bootstrap.lmerMod}
\alias{case.lme}
\alias{case_bootstrap}
\alias{case_bootstrap.lmerMod}
\title{Cases Bootstrap}
\usage{
\method{case_bootstrap}{lmerMod}(model, fn, B, extra_step = FALSE)

case.lme(model, fn, B, extra_step = FALSE)

case_bootstrap(model, fn, B, extra_step)
}
\arguments{
\item{model}{The model object you wish to bootstrap.}

\item{fn}{A function returning the statistic(s) of interest.}

\item{B}{The number of bootstrap resamples.}

\item{extra_step}{}
}
\value{
list
}
\description{
The Cases Bootstrap samples entire cases to generate the bootstrap.
}
\details{
add details
}
\references{
vanderLeeden:208kv
}

