% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstrap_lme4.R, R/bootstrap_nlme.R, R/generics.R
\name{case_bootstrap.lmerMod}
\alias{case_bootstrap}
\alias{case_bootstrap.lme}
\alias{case_bootstrap.lmerMod}
\title{Cases Bootstrap}
\usage{
\method{case_bootstrap}{lmerMod}(model, fn, B, replace)

\method{case_bootstrap}{lme}(model, fn, B, replace)

case_bootstrap(model, fn, B, replace)
}
\arguments{
\item{model}{The model object you wish to bootstrap.}

\item{fn}{A function returning the statistic(s) of interest.}

\item{B}{The number of bootstrap resamples.}

\item{replace}{A logical vector specifying whether each level of the model
should be resampled in the cases bootsrap. The levels should be specified
from the highest level (largest cluster) of the hierarchy to the lowest
(observation-level); for example for students within a school, specify the
school level first, then the student level.}
}
\value{
list
}
\description{
The Cases Bootstrap samples entire cases to generate the bootstrap.
}
\details{
add details
}
\references{
vanderLeeden:208kv
}
\seealso{
Other bootstrap.functions: \code{\link{cgr_bootstrap}},
  \code{\link{cgr_bootstrap.lme}},
  \code{\link{cgr_bootstrap.lmerMod}};
  \code{\link{parametric_bootstrap}},
  \code{\link{parametric_bootstrap.lme}},
  \code{\link{parametric_bootstrap.lmerMod}};
  \code{\link{reb_bootstrap}},
  \code{\link{reb_bootstrap.lme}},
  \code{\link{reb_bootstrap.lmerMod}};
  \code{\link{resid_bootstrap}},
  \code{\link{resid_bootstrap.lme}},
  \code{\link{resid_bootstrap.lmerMod}}
}

