% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.lmeresamp}
\alias{confint.lmeresamp}
\title{confint}
\usage{
\method{confint}{lmeresamp}(object, method, level)
}
\arguments{
\item{object}{The lmeresamp object for which confidence intervals should be computed.}

\item{method}{The type of confidence intervals that should be executed.}

\item{level}{The level at which the confidence interval should be calculated.}
}
\description{
Performs confidence intervals on an lmeresamp object.
}
\details{
This function is given \code{object, method, level} and uses them to calculate the
confidence interval(s).
}
