% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jsp728}
\alias{jsp728}
\title{Junior school project data}
\format{A data frame with 728 rows and 8 columns:}
\source{
\url{http://www.bristol.ac.uk/cmm/team/hg/msm-3rd-ed/datasets.html}
}
\usage{
jsp728
}
\description{
A data set containing...
}
\keyword{datasets}

