% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstrap_lme4.R, R/bootstrap_nlme.R, R/generics.R
\name{parametric_bootstrap.lmerMod}
\alias{parametric_bootstrap}
\alias{parametric_bootstrap.lme}
\alias{parametric_bootstrap.lmerMod}
\title{Parametric Bootstrap}
\usage{
\method{parametric_bootstrap}{lmerMod}(model, fn, B)

\method{parametric_bootstrap}{lme}(model, fn, B)

parametric_bootstrap(model, fn, B)
}
\arguments{
\item{model}{The model object you wish to bootstrap.}

\item{fn}{A function returning the statistic(s) of interest.}

\item{B}{The number of bootstrap resamples.}
}
\description{
The Parametric Bootstrap is uses the parametrically estimated
distribution function of the data to generate bootstrap samples.
}
\details{
This function extracts the fixed effects, simulates from the model, refits the model
and then returns the results in a list.
}
\references{
Chambers, R. and Chandra, H. (2013) A random effect block bootstrap for
   clustered data. \emph{Journal of Computational and Graphical Statistics},
   \bold{22}, 452--470.

   Van der Leeden, R., Meijer, E. and Busing F. M. (2008) Resampling multilevel
   models. In J. de Leeuw and E. Meijer, editors, \emph{Handbook of
   Multilevel Analysis}, pages 401--433. New York: Springer.
}
\seealso{
Other bootstrap.functions: \code{\link{case_bootstrap}},
  \code{\link{case_bootstrap.lme}},
  \code{\link{case_bootstrap.lmerMod}};
  \code{\link{cgr_bootstrap}},
  \code{\link{cgr_bootstrap.lme}},
  \code{\link{cgr_bootstrap.lmerMod}};
  \code{\link{reb_bootstrap}},
  \code{\link{reb_bootstrap.lme}},
  \code{\link{reb_bootstrap.lmerMod}};
  \code{\link{resid_bootstrap}},
  \code{\link{resid_bootstrap.lme}},
  \code{\link{resid_bootstrap.lmerMod}}
}

