% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstrap_lme4.R, R/bootstrap_nlme.R, R/generics.R
\name{parametric_bootstrap.lmerMod}
\alias{parametric_bootstrap}
\alias{parametric_bootstrap.lme}
\alias{parametric_bootstrap.lmerMod}
\title{Parametric Bootstrap}
\usage{
\method{parametric_bootstrap}{lmerMod}(model, fn, B)

\method{parametric_bootstrap}{lme}(model, fn, B)

parametric_bootstrap(model, fn, B)
}
\arguments{
\item{model}{The model object you wish to bootstrap.}

\item{fn}{A function returning the statistic(s) of interest.}

\item{B}{The number of bootstrap resamples.}
}
\value{
list
}
\description{
The Parametric Bootstrap is uses the parametrically estimated
distribution function of the data to generate bootstrap samples.
}
\details{
This function extracts the fixed effects, simulates from the model, refits the model
and then returns the results in a list.
}
\references{
Chambers:2013ba
vanderLeeden:208kv
}

