% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.lmeresamp}
\alias{plot.lmeresamp}
\title{plot}
\usage{
\method{plot}{lmeresamp}(obj, var, ...)
}
\arguments{
\item{obj}{The lmeresamp object to plot.}

\item{var}{The estimated parameter to plot, as a string.}

\item{...}{not used}
}
\description{
Plots the bootstrap call.
}
\details{
This function is given \code{x, var} and uses them to plot the density plot
of bootstrap estimates for the var of choice.
}
