% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.lmeresamp}
\alias{plot.lmeresamp}
\title{Plot bootstrap results}
\usage{
\method{plot}{lmeresamp}(x, var, ...)
}
\arguments{
\item{x}{The lmeresamp object to plot.}

\item{var}{The estimated parameter to plot, as a string or column number.}

\item{...}{not used}
}
\description{
Generate a density plot with a half-eye plot representing the 68% and 95%
percentile intervals from an \code{lmeresamp} object.
}
