% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.lmeresamp}
\alias{print.lmeresamp}
\title{print}
\usage{
\method{print}{lmeresamp}(x, ci = FALSE)
}
\arguments{
\item{x}{The lmeresamp object to print.}

\item{ci}{A logical value specifying whether confidence intervals should be printed.}
}
\description{
Prints the output of the bootstrap call.
}
\details{
This function is given \code{x, ci} and uses them to print the bootstrap result
and confidence intervals if the user wishes to have them.
}
