% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstrap_lme4.R, R/generics.R
\name{reb_bootstrap.lmerMod}
\alias{reb_bootstrap}
\alias{reb_bootstrap.lmerMod}
\title{REB Bootstrap}
\usage{
\method{reb_bootstrap}{lmerMod}(model, fn, B, reb_type = 0)

reb_bootstrap(model, fn, B, reb_type)
}
\arguments{
\item{model}{The model object you wish to bootstrap.}

\item{fn}{A function returning the statistic(s) of interest.}

\item{B}{The number of bootstrap resamples.}

\item{reb_type}{Chooses the type of REB bootstrap}
}
\value{
list
}
\description{
REB Bootstrap
}
\details{
add details
}
\references{
Chambers, R. and Chandra, H. (2013)
A Random Effect Block Bootstrap for Clustered Data.
\emph{Journal of Computational and Graphical Statistics},
22(2), 452–470. doi:10.1080/10618600.2012.681216
}

