% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstrap_lme4.R, R/bootstrap_nlme.R, R/generics.R
\name{reb_bootstrap.lmerMod}
\alias{reb_bootstrap}
\alias{reb_bootstrap.lme}
\alias{reb_bootstrap.lmerMod}
\title{REB Bootstrap}
\usage{
\method{reb_bootstrap}{lmerMod}(model, fn, B, reb_type = 0)

\method{reb_bootstrap}{lme}(model, fn, B, reb_type = 0)

reb_bootstrap(model, fn, B, reb_type)
}
\arguments{
\item{model}{The model object you wish to bootstrap.}

\item{fn}{A function returning the statistic(s) of interest.}

\item{B}{The number of bootstrap resamples.}

\item{reb_type}{Specification of what random effect block bootstrap version to
implement. Possible values are \code{0}, \code{1} or \code{2}.}
}
\value{
list
}
\description{
REB Bootstrap
}
\details{
add details
}
\references{
Chambers, R. and Chandra, H. (2013)
A Random Effect Block Bootstrap for Clustered Data.
\emph{Journal of Computational and Graphical Statistics},
22(2), 452–470. doi:10.1080/10618600.2012.681216
}
\seealso{
Other bootstrap.functions: \code{\link{case_bootstrap}},
  \code{\link{case_bootstrap.lme}},
  \code{\link{case_bootstrap.lmerMod}};
  \code{\link{cgr_bootstrap}},
  \code{\link{cgr_bootstrap.lme}},
  \code{\link{cgr_bootstrap.lmerMod}};
  \code{\link{parametric_bootstrap}},
  \code{\link{parametric_bootstrap.lme}},
  \code{\link{parametric_bootstrap.lmerMod}};
  \code{\link{resid_bootstrap}},
  \code{\link{resid_bootstrap.lme}},
  \code{\link{resid_bootstrap.lmerMod}}
}

