% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstrap_lme4.R, R/bootstrap_nlme.R, R/generics.R
\name{resid_bootstrap.lmerMod}
\alias{resid_bootstrap}
\alias{resid_bootstrap.lme}
\alias{resid_bootstrap.lmerMod}
\title{Residual Bootstrap}
\usage{
\method{resid_bootstrap}{lmerMod}(model, fn, B)

\method{resid_bootstrap}{lme}(model, fn, B)

resid_bootstrap(model, fn, B)
}
\value{
list
}
\description{
The Residual Bootstrap uses residuals to generate bootstrap samples.
}
\details{
Resamples the residuals and complete the bootstrap process.
}
\references{
vanderLeeden:208kv
}

