% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.lmeresamp}
\alias{summary.lmeresamp}
\title{summary}
\usage{
\method{summary}{lmeresamp}(object, ...)
}
\arguments{
\item{object}{The lmeresamp object to be summarized.}

\item{...}{not used}
}
\description{
Summarizes the output of the bootstrap call.
}
\details{
This function is given \code{object} and uses them to summarize the bootstrap result.
}
